Feat/huesped paginado (#24)
* fix:huesped repository

* fix:paginacion de huespedes

+++ src/application/use-cases/huesped/delete-huesped.use-case.ts
@@ -1,9 +1,16 @@
 import { IHuespedRepository } from "../../../domain/interfaces/huesped.repository.interface";
+import { HuespedException } from "../../../domain/exceptions/huesped.exception";
 
 export class DeleteHuespedUseCase {
   constructor(private readonly repository: IHuespedRepository) {}
 
   async execute(id: string): Promise<void> {
+    const existing = await this.repository.findById(id);
+
+    if (!existing) {
+      throw HuespedException.notFoundById(id);
+    }
+
     await this.repository.delete(id);
   }
 }
